feat: Permissions based access to battery pass data

+++ cloud/api.py
@@ -2,24 +2,24 @@ import json
 import os
 import logging
 import pathlib
+
 from datetime import datetime
 from json import JSONDecodeError
-from typing import Literal, Annotated
-
 from Crypto.PublicKey import ECC
 from functools import lru_cache
 from fastapi import FastAPI, Depends, Path, Body
-from fastapi.openapi.models import Example
+from fastapi.params import Query
 from fastapi.responses import JSONResponse
+from pydantic import ValidationError
 from tinydb import TinyDB, where
-from tinydb.table import Document
 
 from crypto.crypto import load_private_key, generate_keys, encrypt_hpke, determine_role, \
     decrypt_hpke, verify_vc
 from dotenv import load_dotenv
-from util.models import EncryptedPayload, SuccessfulResponse, EncryptedPayloadVC, DID, \
-    BadRequestResponse, ForbiddenResponse, NotFoundResponse, VerifiableCredential
-from util.middleware import verify_request
+from util.models import EncryptedPayload, SuccessfulResponse, DID, \
+    BadRequestResponse, ForbiddenResponse, NotFoundResponse, VerifiablePresentation, get_encrypted_payload
+from util.middleware import verify_request, retrieve_data
+from util.validators import validate_battery_pass_payload
 
 app = FastAPI(
     title="Battery Pass API",
@@ -36,18 +36,6 @@ logging.basicConfig(
 )
 with open(pathlib.Path(__file__).parent / "docs" / "example" / "batterypass.json") as f:
     batterypass_json = json.load(f)
-with open(pathlib.Path(__file__).parent / "docs" / "example" / "payload.json") as f:
-    payload_json = json.load(f)
-with open(pathlib.Path(__file__).parent.parent /
-          "blockchain" / "docs" / "VC-DID-examples" / "VC-ServiceAccess.json") as f:
-    vc_service_json = json.load(f)
-
-
-def example_payload(did: DID, vc: VerifiableCredential):
-    payload = payload_json.copy()
-    payload["did"] = did
-    payload["vc"] = vc
-    return payload
 
 
 def error_response(status_code: int, message: str):
@@ -78,7 +66,21 @@ def set_nested_value(doc, path_keys, new_value):
     current_level = doc
     for key in path_keys[:-1]:
         current_level = current_level.setdefault(key, {})
-    current_level[path_keys[-1]] = new_value
+    if isinstance(current_level[path_keys[-1]], list):
+        current_level[path_keys[-1]].append(new_value)
+    else:
+        current_level[path_keys[-1]] = new_value
+
+
+def is_vp(b: bytes) -> VerifiablePresentation | None:
+    try:
+        model = json.loads(b)
+    except (json.JSONDecodeError, UnicodeDecodeError):
+        return None
+    try:
+        return VerifiablePresentation.model_validate(model)
+    except ValidationError:
+        raise ValueError("Invalid Verifiable Presentation.")
 
 
 @app.get("/",
@@ -103,16 +105,6 @@ async def list_dids(db: TinyDB = Depends(get_db)):
     return [entry["did"] for entry in entries if "did" in entry]
 
 
-def retrieve_data(scope: Literal["public", "bms", "legitimate_interest"], did: str, doc: Document,
-                  private_key: ECC.EccKey):
-    if scope not in ["public", "bms", "legitimate_interest"]:
-        raise ValueError(f"Scope '{scope}' is not in ['public', 'bms', 'legitimate_interest'].")
-    decrypted_data = decrypt_hpke(
-        private_key=private_key,
-        bundle=doc["encrypted_data"]
-    )
-
-
 @app.get("/batterypass/{did}",
          summary="Get a battery pass entry by DID",
          tags=["Battery Pass"],
@@ -120,45 +112,43 @@ def retrieve_data(scope: Literal["public", "bms", "legitimate_interest"], did: s
              200: {"model": dict, "content": {"application/json": {"example": batterypass_json}}},
              400: {"model": BadRequestResponse},
              404: {"model": NotFoundResponse},
-         })
+         },
+         description="A detailed description can be found "
+                     "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                     "/blob/main/cloud/docs/api.md#get-batterypassdid)**."
+         )
 async def read_item(
-        did: DID,
-        item: Annotated[EncryptedPayloadVC, Body(openapi_examples={
-            "public": {
-                "summary": "Public scope",
-                "description": "Requires an empty body.",
-                "value": {}
-            },
-            "bms": {
-                "summary": "BMS scope",
-                "description": "Requires a signature signed by the BMS.<br><br>"
-                               "The payload inside the ciphertext must be a **128-byte random number**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=None)
-            },
-            "vc": {
-                "summary": "VC-defined scope",
-                "description": "Requires a verifiable credential signed by the BMS defining the access level.<br><br>"
-                               "The payload inside the ciphertext must be a **128-byte random number**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=vc_service_json)
-            }
-        })],
+        did: str,
+        payload: str = Query(
+            default=None,
+            description="An [encrypted JSON payload](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                        "/blob/main/cloud/docs/api.md#request-body) as a serialized string.\n\n"
+                        "The payload inside the ciphertext can contain a 128-byte random number "
+                        "**or** a [Verifiable Presentation](https://www.w3.org/TR/vc-data-model-2.0/) granting access."
+        ),
+        public: bool = Query(default=True, description="Whether to retrieve only public battery pass data.\n\n"
+                                                       "If set to `false`, `payload` must be provided."),
         db: TinyDB = Depends(get_db),
         private_key: ECC.EccKey = Depends(get_private_key),
 ):
     document = db.search(where("did") == did)
     if not document:
         return error_response(404, "Entry doesn't exist.")
-    if not item:
+    if public:
         return retrieve_data(scope="public", did=did, doc=document[0], private_key=private_key)
+    if not payload:
+        return error_response(400, "Payload must be provided.")
     try:
-        random_number = verify_request(item, private_key)
-        if len(random_number) != 128:
+        payload: EncryptedPayload = EncryptedPayload.model_validate_json(payload) if payload else None
+        decrypted_payload = verify_request(payload, private_key)
+        vp: VerifiablePresentation = is_vp(decrypted_payload)
+        if not vp and len(decrypted_payload) != 128:
             raise ValueError("Invalid length for random value.")
     except ValueError as e:
-        return error_response(400, e.args[0])
-    if determine_role(db, did, item["did"]) == "bms":
+        return error_response(400, str(e))
+    if determine_role(document[0], payload.did) == "bms":
         return retrieve_data(scope="bms", did=did, doc=document[0], private_key=private_key)
-    if "vc" in item and verify_vc(item["vc"]):
+    if vp and verify_vc(vp):
         return retrieve_data(scope="legitimate_interest", did=did, doc=document[0], private_key=private_key)
     return error_response(400, "Invalid request.")
 
@@ -175,33 +165,31 @@ async def read_item(
                    }},
              400: {"model": BadRequestResponse},
              403: {"model": ForbiddenResponse},
-         })
+         },
+         description="A detailed description can be found "
+                     "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                     "/blob/main/cloud/docs/api.md#put-batterypassdid)**.")
 async def create_item(
-        item: Annotated[EncryptedPayload, Body(openapi_examples={
-            "default": {
-                "summary": "Default",
-                "description": "A detailed description can be found "
-                               "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data/blob/main/cloud/docs/api.md#put-batterypassdid)**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=None)
-            }
-        })],
-        did: str = Path(description="A properly formed DID"),
+        payload: EncryptedPayload,
+        did: str,
         db: TinyDB = Depends(get_db),
         private_key: ECC.EccKey = Depends(get_private_key),
 ):
     try:
-        decrypted_item = verify_request(item, private_key)
+        decrypted_payload = verify_request(payload, private_key)
     except ValueError as e:
-        return error_response(400, e.args[0])
-
-    if not determine_role(db, did, item["did"]) == "oem":
-        return error_response(403, "Access denied.")
+        return error_response(400, str(e))
     if db.search(where("did") == did):
         logging.warning(f"DID {did} already exists in DB")
         return error_response(400, "Entry already exists.")
+    if not determine_role(None, payload.did) == "oem":
+        return error_response(403, "Access denied.")
+    results = validate_battery_pass_payload(json.loads(decrypted_payload))
+    if not all(value == "Valid" for value in results.values()):
+        return error_response(400, f"Invalid payload: {json.dumps(results)}")
     db.insert({
         "did": did,
-        "encrypted_data": encrypt_hpke(private_key, decrypted_item)
+        "encrypted_data": encrypt_hpke(private_key, decrypted_payload)
     })
     return {"ok": f"Entry for {did} added successfully."}
 
@@ -219,34 +207,30 @@ async def create_item(
               400: {"model": BadRequestResponse},
               403: {"model": ForbiddenResponse},
               404: {"model": NotFoundResponse},
-          })
+          },
+          description="A detailed description can be found "
+                      "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                      "/blob/main/cloud/docs/api.md#post-batterypassdid)**.")
 async def update_item(
-        item: Annotated[EncryptedPayload, Body(openapi_examples={
-            "default": {
-                "summary": "Default",
-                "description": "A detailed description can be found "
-                               "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data/blob/main/cloud/docs/api.md#post-batterypassdid)**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=None)
-            }
-        })],
-        did: str = Path(description="A properly formed DID"),
+        payload: EncryptedPayload,
+        did: str,
         db: TinyDB = Depends(get_db),
         private_key: ECC.EccKey = Depends(get_private_key),
 ):
     try:
-        decrypted_item = json.loads(verify_request(item, private_key))
+        decrypted_payload = json.loads(verify_request(payload, private_key))
     except JSONDecodeError:
         return error_response(400, "Error occurred while decoding JSON.")
     except ValueError as e:
-        return error_response(400, e.args[0])
+        return error_response(400, str(e))
 
     document = db.search(where("did") == did)
-    if determine_role(db, did, item["did"]) != "bms":
-        return error_response(403, "Access denied.")
     if not document:
         return error_response(404, "Entry doesn't exist.")
+    if determine_role(document[0], payload.did) != "bms":
+        return error_response(403, "Access denied.")
     decrypted_document = json.loads(decrypt_hpke(private_key, document[0]["encrypted_data"]))
-    for element in decrypted_item:  # Iterate over the list of JSON items
+    for element in decrypted_payload:  # Iterate over the list of JSON items
         if not isinstance(element, dict) or len(element) != 1:
             return error_response(400, "Invalid update format.")
         key, value = next(iter(element.items()))
@@ -271,27 +255,25 @@ async def update_item(
                 400: {"model": BadRequestResponse},
                 403: {"model": ForbiddenResponse},
                 404: {"model": NotFoundResponse},
-            })
+            },
+            description="A detailed description can be found "
+                        "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                        "/blob/main/cloud/docs/api.md#delete-batterypassdid)**.")
 async def delete_item(
-        item: Annotated[EncryptedPayload, Body(openapi_examples={
-            "default": {
-                "summary": "Default",
-                "description": "A detailed description can be found "
-                               "**[here](https://github.com/THI-CSI/decentralized_iam_battery_data/blob/main/cloud/docs/api.md#delete-batterypassdid)**.",
-                "value": example_payload(did="did:batterypass:bms.sn-987654321", vc=None)
-            }
-        })],
-        did: str = Path(description="A properly formed DID"),
+        did: str,
+        payload: str = Query(
+            description="An [encrypted JSON payload](https://github.com/THI-CSI/decentralized_iam_battery_data"
+                        "/blob/main/cloud/docs/api.md#request-body) as a serialized string.\n\n"
+                        "The payload inside the ciphertext must contain a 128-byte random number."
+        ),
         db: TinyDB = Depends(get_db),
         private_key: ECC.EccKey = Depends(get_private_key),
 ):
     try:
-        verify_request(item, private_key)
+        payload = EncryptedPayload.model_validate_json(payload)
+        verify_request(payload, private_key)
     except ValueError as e:
-        return error_response(400, e.args[0])
-
-    if determine_role(db, did, item["did"]) != "bms":
-        return error_response(403, "Access denied.")
+        return error_response(400, str(e))
 
     # Search for database entry with the given DID
     document = db.search(where("did") == did)
@@ -299,6 +281,9 @@ async def delete_item(
     # If no entry is found, raise an HTTP exception
     if not document:
         return error_response(404, "Entry doesn't exist.")
+
+    if determine_role(document[0], payload.did) != "bms":
+        return error_response(403, "Access denied.")
 
     # Delete the entry from the database
     db.remove(where("did") == did)
